refactor: changes files and components folders structure, naming & header logo import
